% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVAA.R
\name{CVAA}
\alias{CVAA}
\title{Croos-Values Association Analysis}
\usage{
CVAA(formula, data, samples, cores = 4, OR.col, cl.type = "PSOCK")
}
\arguments{
\item{formula}{A formula. "gene" and "compare" are required and used internally.}

\item{data}{A numeric matrix with n rows(samples) and m cols(genes/variables).}

\item{samples}{A data.frame with sample information.}

\item{cores}{How many cores are used.}

\item{cl.type}{The cluster type for parallel computation.}
}
\description{
Croos-Values Association Analysis
}
\examples{
\dontrun{
library(CVAA)
load(system.file("brca.samples.rda", package="CVAA"))
load(system.file("brca.counts.rda", package="CVAA"))
res <- CVAA(
formula = formula("~(compare+sample_type)*cohort*gene"), 
data = brca.counts, 
samples=brca.samples, 
OR.col = "sample_type", 
cores = 10) 

}

}
